+++ tasks/zagryadskov_m_allreduce/tests/performance/main.cpp
@@ -4,7 +4,6 @@
 #include <cmath>
 #include <cstddef>
 #include <random>
-#include <string>
 #include <vector>
 
 #include "util/include/perf_test_util.hpp"
@@ -29,7 +28,7 @@ class ZagryadskovMRunPerfTestAllreduce : public ppc::util::BaseRunPerfTests<InTy
     std::get<1>(input_data_) = count;
     std::get<2>(input_data_) = op;
     auto &data_vec = std::get<0>(input_data_);
-    data_vec.resize(static_cast<size_t>(count * processes));
+    data_vec.resize(static_cast<size_t>(count) * static_cast<size_t>(processes));
 
     for (int i = 0; i < count * processes; ++i) {
       data_vec[i] = gen(e);
